requirements.txt
absl-py==1.0.0
appdirs==1.4.4
astunparse==1.6.3
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
btrfsutil==5.17
CacheControl==0.12.6
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==37.0.0
cycler==0.11.0
DateTime==4.4
distlib==0.3.4
distro==1.7.0
exchange-calendars==3.6.2
filelock==3.4.2
finnhub-python==2.4.13
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
h5py==3.6.0
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
korean-lunar-calendar==0.2.1
lensfun==0.3.3
libclang==14.0.1
lxml==4.8.0
Markdown==3.3.7
matplotlib==3.5.2
more-itertools==8.12.0
msgpack==1.0.3
multitasking==0.0.10
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==21.3
pandas==1.4.2
pandas-datareader==0.10.0
pandas-market-calendars==3.4
pep517==0.12.0
Pillow==9.1.0
pipenv==2022.5.2
platformdirs==2.5.0
ply==3.11
progress==1.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
pyluach==1.4.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
pytz==2022.1
pywal==3.3.0
requests==2.27.1
requests-oauthlib==1.3.1
resolvelib==0.5.5
retrying==1.3.3
rsa==4.8
scikit-learn==1.1.0
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
team==1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
typing_extensions==4.2.0
urllib3==1.26.9
virtualenv==20.11.0
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.1.2
wikipedia==1.4.0
wrapt==1.14.1
yfinance==0.1.70
zipp==3.8.0
zope.interface==5.4.0
